Ahora la ejecución de las tareas Asincronas se realizada de acuerdo a la version de Android
Starting with Ice Cream Sandwich, the default behavior of AsyncTask has
changed from a parallelized executor to a serialized one.

As you are executing several network requests in a batch of AsyncTask
(as seen in comment), that means your application waits for the previous
request response before lauching the next one.

You can change the executor of an AsyncTask using this code:

if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
    myTask.executeOnExecutor(AsyncTask.THREAD_POOL_EXECUTOR);
}
else {
    myTask.execute();
}